backend/requirements.txt
Flask
Flask-CORS
pymongo
openai==0.28
google-auth
google-auth-oauthlib
google-api-python-client
flask-jwt-extended
python-dotenv
